fix: set `correct_key` after second PageDown input
There was an infinite while loop after the user
input PageDown for the second time, as I forgot to
set the `correct_key` to True. Also, the status
indicator for recording was ending prematurely as
the context manager was in the wrong spot, so I
moved it to the correct location.