feat: try camera settings multiple times
Sometimes when trying to set one of the camera's settings you will get
an error response. I noticed that the next time I connected the camera
and tried to adjust the setting it would work though. I just implemented
a loop to try setting the setting until success or the max number of
retries is reached.