feat: prompt for retry if cam connect fails
When trying to connect to a specific camera the open gopro API will
raise an exception if it can't find the specified camera. This can
happen by bad luck, if it keeps discovering other gopros that are in the
vicinity. If `WirelessGoPro.open()`, either because the specified gopro
could not be found, or because the connection failed, prompt the user
to retry the process. It seems like eventually it will work, if you just
keep retrying.