fix: ESC to cancel recording didn't cancel
The use can press ESC to cancel the recording, provided the recording
has not already started. ESC was not being hanlded correctly and the
application was moving into the "stop recording" context manager, and
the only way to exit that context manager is enter PageDown. Now the ESC
input bypasses the second context manager, as it should.